requirements.txt
Jinja2==3.1.2
pyyaml==6.0.1
markdown==3.4.4
python-markdown-math==0.8